backend/requirements.txt
flask==2.1.2
werkzeug==2.1.2
flask-cors
psycopg2-binary==2.9.5